Restructure file structure
