+++ deeps-engine-latest/src/editor/widget/OpenGLWidget.h
@@ -17,8 +17,8 @@ QT_FORWARD_DECLARE_CLASS(QOpenGLShaderProgram)
 QT_FORWARD_DECLARE_CLASS(QOpenGLTexture)
 QT_FORWARD_DECLARE_CLASS(QWidget)
 
-//#include "../Engine/OpenGLRenderer.h"
-#include "../Engine/Renderer.h"
+//#include "../engine/OpenGLRenderer.h"
+#include "src/engine/renderer/Renderer.h"
 
 class OpenGLWidget : public QOpenGLWidget, protected QOpenGLExtraFunctions {
 Q_OBJECT;
